+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [1.1.2-beta.1](https://github.com/Seungwoo321/tailwind-grid-layout/compare/v1.1.2-beta.0...v1.1.2-beta.1) (2026-02-20)
+
+### Bug Fixes
+
+* **storybook:** resolve MDX file URL issue in pnpm ([be00898](https://github.com/Seungwoo321/tailwind-grid-layout/commit/be00898a94e6d017c32acfb5c1041e218d9ec003))
+
+## [1.1.2-beta.0](https://github.com/Seungwoo321/tailwind-grid-layout/compare/v1.1.1...v1.1.2-beta.0) (2026-02-19)
+
+### Documentation
+
+* **storybook:** update story files for v10 compatibility ([4129a3e](https://github.com/Seungwoo321/tailwind-grid-layout/commit/4129a3e2b211bdf24619a404b206e60db0de645d))
+* update demo and storybook URLs ([325c530](https://github.com/Seungwoo321/tailwind-grid-layout/commit/325c530f4e2ac44dafefb48d5423e8eb9768ce2a))
+
 ## [1.1.1](https://github.com/Seungwoo321/tailwind-grid-layout/compare/v1.1.1-beta.1...v1.1.1) (2026-02-19)
 
 ## [1.1.1-beta.1](https://github.com/Seungwoo321/tailwind-grid-layout/compare/v1.1.1-beta.0...v1.1.1-beta.1) (2026-02-19)